+++ yacchack.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
-cd parsing
-ocamlyacc parser.mly
-mv parser.mly /tmp
+if [ -f parsing/parser.mly ]; then
+  cd parsing
+  ocamlyacc parser.mly
+  mv parser.mly /tmp
+fi
